fix: Company members plugin
- fix delete member error (nonce verification failed and typo in sql prepare statement)
- fix logger (need to use global var, delete Logger class)
- little code formatting
- remove "delete.php" "page"
- css tweak (show admin footer)

+++ wordpress/web/app/plugins/decimus-team-member/inc/Controller/MemberController.php
@@ -13,16 +13,14 @@ use Gulacsi\TeamMember\Exception\Database\{
     PermissionsException,
     UpdateRecordException
 };
-use Gulacsi\TeamMember\Log\Logger;
 use Gulacsi\TeamMember\Form\Validate;
-use wpdb;
+
 
 /**
  * CRUD for members
  */
 class MemberController extends Member implements ControllerInterface
 {
-    use Logger;
     use Validate;
 
 
@@ -41,53 +39,64 @@ class MemberController extends Member implements ControllerInterface
      */
     public function form_action(): void
     {
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
 
-        global $id;
 
-        if ( isset($_POST) && !empty($_POST) ) {
-            $action = $_POST['action'];
+        if (current_user_can('administrator')) {
 
-            if ( $_POST['id'] ?? 0 ) {
-                $id = absint(intval($_POST['id']));
+            global $id;
+
+            if (isset($_POST) && !empty($_POST)) {
+                $action = $_POST['action'];
+
+                if ($_POST['id'] ?? 0) {
+                    $id = absint(intval($_POST['id']));
+                }
+
+                switch ($action) {
+                    // add new member form page
+                    case 'insert':
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/insert.php';
+                        break;
+
+                    // edit member form page
+                    case 'edit':
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/edit.php';
+                        break;
+
+                    // handler function when updating
+                    case 'handle_update':
+                        $this->handle_update();
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/list.php';
+                        break;
+
+                    // handler function when deleting
+                    case 'handle_delete':
+                        $this->handle_delete();
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/list.php';
+                        break;
+
+                    // handler function when inserting new member
+                    case 'handle_insert':
+                        $this->handle_insert();
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/list.php';
+                        break;
+
+                    default:
+                        // list elements
+                        include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/list.php';
+                        break;
+                }
+            } else {
+                include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'pages/member/list.php';
             }
-
-            switch ($action) {
-                // add new member form page
-                case 'insert':
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/insert.php';
-                    break;
-
-                // edit member form page
-                case 'edit':
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/edit.php';
-                    break;
-
-                // handler function when updating
-                case 'handle_update':
-                    $this->handle_update();
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/list.php';
-                    break;
-
-                // handler function when deleting
-                case 'handle_delete':
-                    $this->handle_delete();
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/list.php';
-                    break;
-
-                // handler function when inserting new member
-                case 'handle_insert':
-                    $this->handle_insert();
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/list.php';
-                    break;
-
-                default:
-                    // list elements
-                    include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/list.php';
-                    break;
-            }
-        } else {
-            include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . 'pages/member/list.php';
         }
     }
 
@@ -98,25 +107,46 @@ class MemberController extends Member implements ControllerInterface
      */
     public function handle_insert(): void
     {
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
 
-        // !!! verify insert nonce !!!
-        $this->verify_nonce('insert');
 
         try {
+            // !!! verify insert nonce !!!
+            $this->verify_nonce('insert');
+
+            if (!current_user_can('manage_options')) {
+                throw new PermissionsException('You do not have sufficient permissions to view this page.');
+            }
+
+
             // get sanitized form values from inputs
             $sanitized_data = $this->get_form_input_field_values();
 
             $response = $this->insert($sanitized_data);
 
-            if ( $response === false ) {
+            if ($response === false) {
                 throw new InsertRecordException('Unable to insert new record.');
             } else {
                 echo '<div class="notice notice-success is-dismissible"><p>Team member successfully added.</p></div>';
             }
-        } catch (InsertRecordException|Exception $ex) {
-            echo '<div class="notice notice-error"><p>' . $ex->getMessage() . '</p></div>';
-            $this->exception_logger(self::LOGGING, $ex);
+
+        } catch (InsertRecordException|PermissionsException|Exception $ex) {
+
+            echo '<div class="notice notice-error"><p>'.$ex->getMessage().'</p></div>';
+
+            if (self::LOGGING) {
+                global $decimus_team_member_log;
+                $decimus_team_member_log->logError(
+                    $ex->getMessage()." - ".__FILE__.":".__METHOD__.":".__LINE__);
+            }
+
         }
 
     }
@@ -129,12 +159,25 @@ class MemberController extends Member implements ControllerInterface
     public function handle_update(): void
     {
         // debug log and log to file
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
 
-        // !!! verify edit nonce !!!
-        $this->verify_nonce('edit');
 
         try {
+            // !!! verify edit nonce !!!
+            $this->verify_nonce('edit');
+
+            if (!current_user_can('manage_options')) {
+                throw new PermissionsException('You do not have sufficient permissions to view this page.');
+            }
+
+
             // get sanitized form values from inputs
             $data = $this->get_form_input_field_values();
 
@@ -146,14 +189,22 @@ class MemberController extends Member implements ControllerInterface
                 $response = $this->update($data, true);
             }
 
-            if ( $response === false ) {
+            if ($response === false) {
                 throw new UpdateRecordException('Unable to update member record.');
             } else {
                 echo '<div class="notice notice-success is-dismissible"><p>Team member data successfully updated.</p></div>';
             }
-        } catch (UpdateRecordException|Exception $ex) {
-            echo '<div class="notice notice-error"><p>' . $ex->getMessage() . '</p></div>';
-            $this->exception_logger(self::LOGGING, $ex);
+
+        } catch (UpdateRecordException|PermissionsException|Exception $ex) {
+
+            echo '<div class="notice notice-error"><p>'.$ex->getMessage().'</p></div>';
+
+            if (self::LOGGING) {
+                global $decimus_team_member_log;
+                $decimus_team_member_log->logError(
+                    $ex->getMessage()." - ".__FILE__.":".__METHOD__.":".__LINE__);
+            }
+
         }
 
     }
@@ -165,25 +216,46 @@ class MemberController extends Member implements ControllerInterface
      */
     public function handle_delete(): void
     {
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
 
-        $this->verify_nonce('delete');
 
         try {
-            if ( $_POST['id'] ?? 0 ) {
+
+            $this->verify_nonce('delete');
+
+            if (!current_user_can('manage_options')) {
+                throw new PermissionsException('You do not have sufficient permissions to view this page.');
+            }
+
+
+            if ($_POST['id'] ?? 0) {
                 $id = absint(intval($_POST['id']));
 
                 $response = $this->delete($id);
 
-                if ( $response === false ) {
+                if ($response === false) {
                     throw new DeleteRecordException('Unable to delete team member.');
                 } else {
                     echo '<div class="notice notice-success is-dismissible"><p>Team member successfully deleted.</p></div>';
                 }
             }
-        } catch (DeleteRecordException|Exception $ex) {
-            echo '<div class="notice notice-error"><p>' . $ex->getMessage() . '</p></div>';
-            $this->exception_logger(self::LOGGING, $ex);
+
+        } catch (DeleteRecordException|PermissionsException|Exception $ex) {
+
+            echo '<div class="notice notice-error"><p>'.$ex->getMessage().'</p></div>';
+            if (self::LOGGING) {
+                global $decimus_team_member_log;
+                $decimus_team_member_log->logError(
+                    $ex->getMessage()." - ".__FILE__.":".__METHOD__.":".__LINE__);
+            }
+
         }
 
     }
@@ -195,21 +267,37 @@ class MemberController extends Member implements ControllerInterface
      */
     public function add_form(): void
     {
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
+
 
         try {
-            if ( !current_user_can('manage_options') ) {
+
+            if (!current_user_can('manage_options')) {
                 throw new PermissionsException('You do not have sufficient permissions to view this page.');
             }
 
-            if ( !empty($_POST) ) {
+            if (!empty($_POST)) {
                 $this->form_action();
             } else {
-                include DECIMUS_TEAM_MEMBER_PLUGIN_DIR . '/pages/member/insert.php';
+                include DECIMUS_TEAM_MEMBER_PLUGIN_DIR.'/pages/member/insert.php';
             }
+
         } catch (PermissionsException|Exception $ex) {
-            echo '<div class="notice notice-error"><p>' . $ex->getMessage() . '</p></div>';
-            $this->exception_logger(self::LOGGING, $ex);
+
+            echo '<div class="notice notice-error"><p>'.$ex->getMessage().'</p></div>';
+            if (self::LOGGING) {
+                global $decimus_team_member_log;
+                $decimus_team_member_log->logError(
+                    $ex->getMessage()." - ".__FILE__.":".__METHOD__.":".__LINE__);
+            }
+
         }
     }
 
@@ -220,24 +308,38 @@ class MemberController extends Member implements ControllerInterface
      */
     public function list_table(): void
     {
-        $this->logger(self::DEBUG, self::LOGGING);
+        if (self::LOGGING) {
+            global $decimus_team_member_log;
+            $decimus_team_member_log->logInfo("Entering - ".__FILE__.":".__METHOD__.":".__LINE__);
+        }
+        if (self::DEBUG) {
+            $info_text = "Entering - ".__FILE__.":".__METHOD__.":".__LINE__;
+            echo '<div class="notice notice-info is-dismissible">'.$info_text.'</p></div>';
+        }
+
 
         try {
-            /** @noinspection PhpUnusedLocalVariableInspection */
-            /** @var wpdb $wpdb */
+
             global $wpdb;
 
             // note: current_user_can() always returns false if the user is not logged in
-            if ( !current_user_can('manage_options') ) {
+            if (!current_user_can('manage_options')) {
                 throw new PermissionsException(
                     'You do not have sufficient permissions to view this page.'
                 );
             }
+
             $this->form_action();
 
         } catch (PermissionsException|Exception $ex) {
-            echo '<div class="notice notice-error"><p>' . $ex->getMessage() . '</p></div>';
-            $this->exception_logger(self::LOGGING, $ex);
+
+            echo '<div class="notice notice-error"><p>'.$ex->getMessage().'</p></div>';
+            if (self::LOGGING) {
+                global $decimus_team_member_log;
+                $decimus_team_member_log->logError(
+                    $ex->getMessage()." - ".__FILE__.":".__METHOD__.":".__LINE__);
+            }
+
         }
     }
